fix(mypy): fix duplicate module errors
