+++ internal/factories/builder_factory.py
@@ -1,8 +1,8 @@
 '''
-@File     :  generator_factory.py
-@Desc     :  Generator factory, to create the correct generator based on the report mode
+@File     :  builder_factory.py
+@Desc     :  Builder factory, to create the correct builder based on the report mode
 @Authors  :  Nick Vissers <nick.vissers@openpixelsystems.org>
-@Date     :  25/12/2023
+@Date     :  24/12/2023
 @Version  :  1.0
 @License  :  Copyright (C) 2023 Open Pixel Systems. All rights reserved.
           :  Confidential and for internal use only. The content of this document constitutes proprietary
@@ -10,25 +10,24 @@
           :  of any part of the content of this document by unauthorized parties is strictly prohibited.
 '''
 
+from internal.builder import Builder
 from internal.collector import Asset
-from internal.dynamic.generator import DynamicGenerator
-from internal.generator import Generator
+from internal.dynamic.dynamic_builder import DynamicBuilder
 from internal.logger import Logger
-from internal.static.generator import StaticGenerator
-
+from internal.static.static_builder import StaticBuilder
 from internal.internal_types import ReportMode
 
 
-class GeneratorFactory:
+class BuilderFactory:
     class Error(Exception):
         def __init__(self):
             super().__init__("Unknown report mode")
 
     @staticmethod
-    def create(mode: ReportMode, styles: list[Asset], scripts: list[Asset], logger: Logger) -> Generator:
+    def create(mode: ReportMode, styles: list[Asset], scripts: list[Asset], logger: Logger) -> Builder:
         if mode == ReportMode.STATIC:
-            return StaticGenerator(styles, scripts, logger)
+            return StaticBuilder(styles, scripts, logger)
         elif mode == ReportMode.DYNAMIC:
-            return DynamicGenerator(styles, scripts, logger)
+            return DynamicBuilder(styles, scripts, logger)
 
-        raise GeneratorFactory.Error()
+        raise BuilderFactory.Error()
