+++ mypy.ini
@@ -1,0 +1,3 @@
+[mypy]
+
+mypy_path = internal
